+++ monew/src/main/java/com/spring/monew/batch/exception/BatchJobExceptionHandler.java
@@ -1,4 +1,4 @@
-package com.spring.monew.article.batch.exception;
+package com.spring.monew.batch.exception;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
