+++ monew/src/main/java/com/spring/monew/batch/exception/BatchJobExecutionNotFoundException.java
@@ -1,4 +1,4 @@
-package com.spring.monew.article.batch.exception;
+package com.spring.monew.batch.exception;
 
 public class BatchJobExecutionNotFoundException extends RuntimeException {
     public BatchJobExecutionNotFoundException(String message) {
