+++ monew/src/main/resources/schema-h2.sql
@@ -136,3 +136,104 @@ CREATE TABLE notifications
         REFERENCES users (id)
         ON DELETE CASCADE ON UPDATE CASCADE
 );
+
+
+
+-- Spring Batch 5.x 용 PostgreSQL 메타데이터 DDL
+-- 3개의 ID 생성용 시퀀스 (가장 중요한 누락 부분)
+CREATE SEQUENCE BATCH_JOB_SEQ MAXVALUE 9223372036854775807 NO CYCLE;
+CREATE SEQUENCE BATCH_JOB_EXECUTION_SEQ MAXVALUE 9223372036854775807 NO CYCLE;
+CREATE SEQUENCE BATCH_STEP_EXECUTION_SEQ MAXVALUE 9223372036854775807 NO CYCLE;
+
+
+-- Job Instance
+CREATE TABLE batch_job_instance (
+                                    job_instance_id BIGINT NOT NULL,
+                                    version BIGINT,
+                                    job_name VARCHAR(100) NOT NULL,
+                                    job_key VARCHAR(32) NOT NULL,
+                                    CONSTRAINT batch_job_instance_pkey PRIMARY KEY (job_instance_id),
+                                    CONSTRAINT job_inst_un UNIQUE (job_name, job_key)
+);
+
+-- Job Execution
+CREATE TABLE batch_job_execution (
+                                     job_execution_id BIGINT NOT NULL,
+                                     version BIGINT,
+                                     job_instance_id BIGINT NOT NULL,
+                                     create_time TIMESTAMP WITHOUT TIME ZONE NOT NULL,
+                                     start_time TIMESTAMP WITHOUT TIME ZONE,
+                                     end_time TIMESTAMP WITHOUT TIME ZONE,
+                                     status VARCHAR(10),
+                                     exit_code VARCHAR(2500),
+                                     exit_message VARCHAR(2500),
+                                     last_updated TIMESTAMP WITHOUT TIME ZONE,
+                                     job_configuration_location VARCHAR(2500) NULL,
+                                     CONSTRAINT batch_job_execution_pkey PRIMARY KEY (job_execution_id)
+);
+
+-- Step Execution
+CREATE TABLE batch_step_execution (
+                                      step_execution_id BIGINT NOT NULL,
+                                      version BIGINT NOT NULL,
+                                      step_name VARCHAR(100) NOT NULL,
+                                      job_execution_id BIGINT NOT NULL,
+                                      create_time TIMESTAMP WITHOUT TIME ZONE NOT NULL, -- \d 에 존재 확인
+                                      start_time TIMESTAMP WITHOUT TIME ZONE,
+                                      end_time TIMESTAMP WITHOUT TIME ZONE,
+                                      status VARCHAR(10),
+                                      commit_count BIGINT,
+                                      read_count BIGINT,
+                                      filter_count BIGINT,
+                                      write_count BIGINT,
+                                      read_skip_count BIGINT,
+                                      write_skip_count BIGINT,
+                                      process_skip_count BIGINT,
+                                      rollback_count BIGINT,
+                                      exit_code VARCHAR(2500),
+                                      exit_message VARCHAR(2500),
+                                      last_updated TIMESTAMP WITHOUT TIME ZONE,
+                                      CONSTRAINT batch_step_execution_pkey PRIMARY KEY (step_execution_id) -- \d 에서 누락되었으나 필수
+);
+
+-- Job Execution Context
+CREATE TABLE batch_job_execution_context (
+                                             job_execution_id BIGINT NOT NULL,
+                                             short_context VARCHAR(2500) NOT NULL,
+                                             serialized_context TEXT,
+                                             CONSTRAINT batch_job_execution_context_pkey PRIMARY KEY (job_execution_id)
+);
+
+-- Step Execution Context
+CREATE TABLE batch_step_execution_context (
+                                              step_execution_id BIGINT NOT NULL,
+                                              short_context VARCHAR(2500) NOT NULL,
+                                              serialized_context TEXT,
+                                              CONSTRAINT batch_step_execution_context_pkey PRIMARY KEY (step_execution_id)
+);
+
+-- Job Execution Parameters (Batch 5.x)
+CREATE TABLE batch_job_execution_params (
+                                            job_execution_id BIGINT NOT NULL,
+                                            parameter_name VARCHAR(100) NOT NULL,
+                                            parameter_type VARCHAR(100) NOT NULL,
+                                            parameter_value VARCHAR(2500),
+                                            identifying CHAR(1) NOT NULL
+);
+
+
+-- 외래 키(Foreign Key) 제약 조건
+ALTER TABLE batch_job_execution
+    ADD CONSTRAINT job_inst_exec_fk FOREIGN KEY (job_instance_id) REFERENCES batch_job_instance(job_instance_id);
+
+ALTER TABLE batch_job_execution_params
+    ADD CONSTRAINT job_exec_params_fk FOREIGN KEY (job_execution_id) REFERENCES batch_job_execution(job_execution_id);
+
+ALTER TABLE batch_step_execution
+    ADD CONSTRAINT job_exec_step_fk FOREIGN KEY (job_execution_id) REFERENCES batch_job_execution(job_execution_id);
+
+ALTER TABLE batch_job_execution_context
+    ADD CONSTRAINT job_exec_ctx_fk FOREIGN KEY (job_execution_id) REFERENCES batch_job_execution(job_execution_id);
+
+ALTER TABLE batch_step_execution_context
+    ADD CONSTRAINT step_exec_ctx_fk FOREIGN KEY (step_execution_id) REFERENCES batch_step_execution(step_execution_id);
